move unneeded_wildcard_pattern to its own module
